gpsX ,gpsY 의 DecimalField를 FloatField로 변경

+++ InterfaceApp/models.py
@@ -9,8 +9,9 @@ class PerformanceList(models.Model):
     realmName = models.CharField(max_length=10)
     area = models.CharField(max_length=10, null=True)
     thumbnail = models.CharField(max_length=256)
-    gpsX = models.DecimalField(null=True)
-    gpsY = models.DecimalField(null=True)
+    gpsX = models.FloatField(null=True)
+    gpsY = models.FloatField(null=True)
+
 
     def __str__(self):
         return f"seq: {self.seq}, title: {self.title}, startDate: {self.startDate}, endDate: {self.endDate}, place: {self.place}, realmName: {self.realmName}, area: {self.area}, thumbnail: {self.thumbnail}, gpsX: {self.gpsX}, gpsY: {self.gpsY}"
